bug: Fix promise issue
Fix promise issue where Jetstream would not wait for Producer to startup before carrying on;
Refactor index.ts to remove the factory from there and rather be its own class;
Rename SERVER_TYPE to STARTUP_TYPE as server type seems to be reserved for NodeJS.
