+++ .env.template.jetstream
@@ -1,9 +1,10 @@
+STARTUP_TYPE='jetstream'
 NODE_ENV=
 SERVER_URL=
 FUNCTION_NAME=
 PRODUCER_STREAM=
 CONSUMER_STREAM=
 STREAM_SUBJECT=
-ACK_POLICY='None' | 'All' | 'Explicit' | 'NotSet'
-PRODUCER_STORAGE=
+ACK_POLICY='Explicit' | 'All'
+PRODUCER_STORAGE='File' | 'Memory'
 PRODUCER_RETENTION_POLICY='Workqueue' | 'Interest' | 'Limits'
