+++ .env.template.nats
@@ -1,3 +1,4 @@
+STARTUP_TYPE='nats'
 NODE_ENV=
 SERVER_URL=
 FUNCTION_NAME=
